added drum patterns to the looper
added drum patterns to the looper - it works, but not the way I was setting it up. The pattern is only 1 measure of 16th notes -  I was shooting for 4 measures of 16th notes for a total of 64. Changes the piano sample set to one online that sounds better - will download and add to the project.